Move-only: Standard namespace with general CWDestination variant and encode/decode destination functions moved to their own file.
